chore: Rename tasks.rs to task.rs
