+++ server/services/manager/src/api/mod.rs
@@ -4,7 +4,7 @@ use crate::AppState;
 
 mod health;
 mod openapi_docs;
-mod tasks;
+mod task;
 mod worker_kind;
 
 pub fn routes() -> Router<AppState> {
@@ -12,5 +12,5 @@ pub fn routes() -> Router<AppState> {
         .nest("/worker-kind", worker_kind::routes())
         .nest("/api-docs", openapi_docs::routes())
         .nest("/health", health::routes())
-        .nest("/tasks", tasks::routes())
+        .nest("/tasks", task::routes())
 }
